Allow /statuses to return a 303 See Other response.
According to: https://www.ietf.org/rfc/rfc2616.txt a 303 See Other can
be returned to indicate that the resource the user is looking for can be
found under a new path.
In the case of a transaction including the `status == 'valid'`, we
return 303 See Other, as well as a Location header to the /transactions
endpoint.

"The response to the request can be found under a different URI and SHOULD be
retrieved using a GET method on that resource."