Update setup.py for BigchainDB 2.0
I took the license name from this list of valid PyPI licenses:

https://jlesquembre.github.io/autopilot/license_list.html

About line 112: The docs will say it's production-ready depending on your use case so I figured we could update that. It's not like this is a contract or anything.

I'm not sure if we should keep line 120:

```python
'Programming Language :: Python :: 3.5',
```
or line 122:
```python
 'Operating System :: MacOS :: MacOS X',
```
It didn't actually work on mac os, not sure why we had that. Does it work there now?