edit BigchainDB license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit adds a LICENSE file that contains the full text of the
Apache 2.0 license. (The existing LICENSES.md file has been left as-is).

This change allows Licensee to successfully identify the license type of
BigchainDB as Apache 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>