Problem: arbitrary big transactions might fail
Solution: add a new configuration parameter `max_transaction_size` to
limit the size of transactions accepted by the web API. The parameter is
also exposed by the root API endpoint using the same key
`max_transaction_size`, so drivers can act accordingly and eventually
fail early if the transaction to be sent is bigger than the upper limit.
This issue was first reported in #2222.

By default Flask will accept file uploads to an unlimited amount of
memory, but you can limit that by setting the [`MAX_CONTENT_LENGTH`
config key][1]. Spoofing the `Content-Length` is also considered: see
the integration test under
`acceptance/python/src/test_max_transaction_size.py`.

I had troubles creating a unit test to check this feature. It seems that
the pytest fixture for testing Flask ignores the `MAX_CONTENT_LENGTH`
setting. If you have better ideas on how to test that, please make a
patch.

To be honest, I'm not sure that having the parameter
`max_transaction_size` in the BigchainDB configuration file is a good
idea. It should be a Network wide configuration parameter, shared and
accepted by every non-byzantine node. We should have something like the
`genesis.json` file in Tendermint, and it should be updated using the
Election Process specified by [BEP-18][2]. If we do that, then
`max_transaction_size` should be checked also in the validation rules.

[1]: http://flask.pocoo.org/docs/1.0/patterns/fileuploads/#improving-uploads
[2]: https://github.com/bigchaindb/BEPs/tree/master/18